requirements.txt
bs4==0.0.1
pygame==1.9.3
pyperclip==1.7.0
requests==2.21.0
websocket-client==0.54.0
